produces results; visualization not working yet

+++ progressbar.py
@@ -9,6 +9,7 @@ Usage:
 
 import terminal_progress as terminal
 import sys
+import numpy as np
 
 class ProgressBar(object):
     """Terminal progress bar class"""
@@ -57,7 +58,7 @@ class ProgressBar(object):
         # Check if render is called for the first time
         if self.progress != None:
             self.clear()
-        self.progress = (bar_width * percent) / 100
+        self.progress = (bar_width * percent.astype(np.int)) / 100
         data = self.TEMPLATE % {
             'percent': percent,
             'color': self.color,
